put general utility functions under banditpylib package
