+++ banditpylib/__init__.py
@@ -1,0 +1,6 @@
+from .utils import *
+
+__all__ = [
+    'argmax_or_min',
+    'argmax_or_min_tuple',
+]
