+++ banditpylib/learners/__init__.py
@@ -1,6 +1,6 @@
 from .utils import *
 
 __all__ = [
-    'argmax_or_min', 'argmax_or_min_tuple', 'Goal', 'BestArmId', 'MaxReward',
-    'MaxCorrectAnswers', 'AllCorrect', 'Learner'
+    'Goal', 'BestArmId', 'MaxReward', 'MaxCorrectAnswers', 'AllCorrect',
+    'Learner'
 ]
