+++ banditpylib/learners/ordinary_fbbai_learner/sr.py
@@ -3,9 +3,9 @@ from typing import Dict
 import math
 import numpy as np
 
+from banditpylib import argmax_or_min_tuple
 from banditpylib.arms import PseudoArm
 from banditpylib.data_pb2 import Actions, Feedback
-from banditpylib.learners import argmax_or_min_tuple
 from .utils import OrdinaryFBBAILearner
