+++ banditpylib/learners/ordinary_fcbai_learner/lilucb_heur.py
@@ -1,9 +1,9 @@
 import math
 import numpy as np
 
+from banditpylib import argmax_or_min_tuple
 from banditpylib.arms import PseudoArm
 from banditpylib.data_pb2 import Actions, Feedback
-from banditpylib.learners import argmax_or_min_tuple
 from .utils import OrdinaryFCBAILearner
